backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-docx==1.1.0
python-pptx==0.6.23
groq==0.36.0
python-dotenv==1.0.0
email-validator==2.3.0
requests==2.31.0
Pillow==10.2.0
